+++ src/components/init/init-info-confirmation/type.ts
@@ -1,0 +1,6 @@
+import { ReactNode } from 'react';
+
+export type CustomButtonProps = {
+  text: string;
+  icon: ReactNode;
+};
